requirements.txt
asgiref==3.10.0
black==25.9.0
click==8.3.0
colorama==0.4.6
Django==5.0.14
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pytokens==0.2.0
sqlparse==0.5.3
tzdata==2025.2
qrcode[pil]==7.3.1
